feat(server): add GET /v1/sync endpoint, remove orphaned routes (#98)

+++ app/src/main/java/eu/monniot/feed/store/RoomArticleStore.kt
@@ -20,7 +20,7 @@ class RoomArticleStore(private val db: RoomDatabase, private val dao: ArticleSto
         dao.upsert(articles.map { it.toEntity() })
     }
 
-    override suspend fun deleteByIds(ids: List<Int>) {
+    override suspend fun deleteByIds(ids: List<Long>) {
         ids.chunked(900).forEach { chunk ->
             dao.deleteByIds(chunk)
         }
